Create spxrepair_header
Header file that will be prefixed into all scripted repair operations. This sets the appropriate environments for the repair and then initiates the call to sp_ctrl. The spxrepair.sh script then appends the individual repair commands to this and closes the EOF. This is a template that is used to populate the repair script only.