Create spxenv.sh
Set up the Shareplex environment. You may need to edit this file according to your needs. We reserve the "spx" identifier in /etc/oratab for Shareplex database SIDs, and "splex" directory naming for Shareplex shared directories, so we can get away with grepping for these values and using a single file for every environment. You can adjust or hardcode as needed in yours.